Rename `webview/src/vscode-api.ts` ti `webview/src/message.ts`
To continue trying to isolate the webview from VS Code specific stuff.
