+++ webview/src/components/FormatTab.tsx
@@ -1,5 +1,5 @@
 import type { CommandResult, FileInfo } from '../../../protocol/types';
-import { vscode } from '../vscode-api';
+import { vscode } from '../message';
 import { RawOutput } from './RawOutput';
 import { Info, CheckCircle, AlertCircle, FileQuestion } from 'lucide-react';
